requirements.txt
flask>=3.1.0
openai>=1.50.0
twilio>=9.0.0
python-dotenv>=1.0.0
requests>=2.32.0
numpy>=2.0.0